Move Common Modifiers and PaletteEffects into Traits space
