+++ OpenRA.Mods.Common/OpenRA.Mods.Common.csproj
@@ -79,40 +79,40 @@
     <Compile Include="Graphics\VoxelRenderable.cs" />
     <Compile Include="LoadWidgetAtGameStart.cs" />
     <Compile Include="ModChooserLoadScreen.cs" />
-    <Compile Include="Modifiers\DisabledOverlay.cs" />
-    <Compile Include="Modifiers\HiddenUnderFog.cs" />
-    <Compile Include="Modifiers\UpgradeOverlay.cs" />
     <Compile Include="NullLoadScreen.cs" />
     <Compile Include="Orders\DeployOrderTargeter.cs" />
     <Compile Include="Orders\EnterAlliedActorTargeter.cs" />
     <Compile Include="Orders\UnitOrderTargeter.cs" />
-    <Compile Include="PaletteEffects\CloakPaletteEffect.cs" />
-    <Compile Include="PaletteEffects\LightPaletteRotator.cs" />
-    <Compile Include="PaletteEffects\MenuPaletteEffect.cs" />
-    <Compile Include="PaletteEffects\NukePaletteEffect.cs" />
-    <Compile Include="PaletteEffects\WaterPaletteRotation.cs" />
     <Compile Include="ServerTraits\ColorValidator.cs" />
     <Compile Include="ServerTraits\LobbyCommands.cs" />
     <Compile Include="ServerTraits\LobbySettingsNotification.cs" />
     <Compile Include="ServerTraits\MasterServerPinger.cs" />
     <Compile Include="ServerTraits\PlayerPinger.cs" />
     <Compile Include="Traits\BlocksBullets.cs" />
-    <Compile Include="Traits\Immobile.cs" />
     <Compile Include="Traits\Burns.cs" />
     <Compile Include="Traits\CustomBuildTimeValue.cs" />
     <Compile Include="Traits\CustomSellValue.cs" />
     <Compile Include="Traits\Demolishable.cs" />
+    <Compile Include="Traits\Immobile.cs" />
     <Compile Include="Traits\JamsMissiles.cs" />
-    <Compile Include="Traits\ProvidesRadar.cs" />
-    <Compile Include="Traits\RallyPoint.cs" />
-    <Compile Include="Traits\ShakeOnDeath.cs" />
+    <Compile Include="Traits\Modifiers\DisabledOverlay.cs" />
+    <Compile Include="Traits\Modifiers\HiddenUnderFog.cs" />
+    <Compile Include="Traits\Modifiers\UpgradeOverlay.cs" />
+    <Compile Include="Traits\PaletteEffects\CloakPaletteEffect.cs" />
+    <Compile Include="Traits\PaletteEffects\LightPaletteRotator.cs" />
+    <Compile Include="Traits\PaletteEffects\MenuPaletteEffect.cs" />
+    <Compile Include="Traits\PaletteEffects\NukePaletteEffect.cs" />
+    <Compile Include="Traits\PaletteEffects\WaterPaletteRotation.cs" />
     <Compile Include="Traits\Power\AffectedByPowerOutage.cs" />
     <Compile Include="Traits\Power\CanPowerDown.cs" />
     <Compile Include="Traits\Power\Player\PowerManager.cs" />
     <Compile Include="Traits\Power\Power.cs" />
     <Compile Include="Traits\Power\RequiresPower.cs" />
     <Compile Include="Traits\Power\ScalePowerWithHealth.cs" />
+    <Compile Include="Traits\ProvidesRadar.cs" />
+    <Compile Include="Traits\RallyPoint.cs" />
     <Compile Include="Traits\Render\RenderSprites.cs" />
+    <Compile Include="Traits\ShakeOnDeath.cs" />
     <Compile Include="Traits\Sound\ActorLostNotification.cs" />
     <Compile Include="Traits\Sound\AnnounceOnBuild.cs" />
     <Compile Include="Traits\Sound\AnnounceOnKill.cs" />
