+++ OpenRA.Mods.Common/Traits/PaletteEffects/CloakPaletteEffect.cs
@@ -13,7 +13,7 @@ using System.Drawing;
 using OpenRA.Graphics;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.Common
+namespace OpenRA.Mods.Common.Traits
 {
 	public class CloakPaletteEffectInfo : TraitInfo<CloakPaletteEffect> { }
