+++ OpenRA.Mods.RA/Widgets/Logic/Ingame/LeaveMapLogic.cs
@@ -11,7 +11,7 @@
 using System;
 using System.Drawing;
 using System.Linq;
-using OpenRA.Mods.Common;
+using OpenRA.Mods.Common.Traits;
 using OpenRA.Mods.RA.Scripting;
 using OpenRA.Network;
 using OpenRA.Traits;
